+++ src/commands/index.ts
@@ -1,3 +1,4 @@
+export { deleteSingleLogCommand, toggleCommentCommand, updateSingleLogCommand } from './codelens-actions';
 export { commentAllLogsCommand } from './comment-logs';
 export { deleteAllLogsCommand } from './delete-logs';
 export { insertLogCommand } from './insert-log';